# Performance improvement of `strictNullChecks`