--- a/src/test/kotlin/com/fasterxml/jackson/module/kotlin/KotlinInstantiatorsTest.kt
+++ b/src/test/kotlin/com/fasterxml/jackson/module/kotlin/KotlinInstantiatorsTest.kt
@@ -13,7 +13,6 @@ class KotlinInstantiatorsTest {
         nullToEmptyCollection = false,
         nullToEmptyMap = false,
         nullIsSameAsDefault = false,
-        strictNullChecks = false
     )
 
     @Test