@k163377 2025-01-13
If the exceptions raised were to be changed destructively, they needed to be deprecated correctly.
It needs to be fundamentally reworked and therefore `drafted`.